Sentry_AI_V2/requirements.txt
# AI / ML
ultralytics==8.2.77
torch
torchvision
decord
pytorchvideo

# Computer vision
opencv-python
Pillow
numpy
pandas

# Telegram alerts
python-telegram-bot==20.3

# PDF generation
fpdf2==2.7.4
reportlab==4.0.0

# Database / logging
SQLAlchemy==2.0.38
